kafka-producer/python/requirements.txt
aiokafka==0.10.0
async-timeout==4.0.3
kafka-python==2.0.2
packaging==23.2
SQLAlchemy==2.0.25
typing_extensions==4.9.0
psycopg2-binary==2.9.9